Fixes breaking post-rebase changes re: ElementType, moving TNS, MNS, if self._subfolders check
