+++ exchangelib/ewsdatetime.py
@@ -102,6 +102,8 @@ class EWSDateTime(datetime.datetime):
 
     @classmethod
     def from_datetime(cls, d):
+        if type(d) == cls:
+            return d
         if type(d) != datetime.datetime:
             raise ValueError("%r must be a datetime instance" % d)
         if d.tzinfo is None:
